re_types_core::component_descriptor
Struct ComponentDescriptor

pub struct ComponentDescriptor {
    pub archetype_name: Option<ArchetypeName>,
    pub archetype_field_name: Option<ArchetypeFieldName>,
    pub component_name: ComponentName,
}

A ComponentDescriptor fully describes the semantics of a column of data.

Every component is uniquely identified by its ComponentDescriptor.

Fields
archetype_name: Option<ArchetypeName>

Optional name of the Archetype associated with this data.

None if the data wasn’t logged through an archetype.

Example: rerun.archetypes.Points3D.

archetype_field_name: Option<ArchetypeFieldName>

Optional name of the field within Archetype associated with this data.

None if the data wasn’t logged through an archetype.

Example: positions.

component_name: ComponentName

Semantic name associated with this data.

This is fully implied by archetype_name and archetype_field, but included for semantic convenience.

Example: rerun.components.Position3D.

Implementations

impl ComponentDescriptor

fn to_any_string(&self, use_short_names: bool) -> String

pub fn sanity_check(&self)

pub fn full_name(&self) -> String

Returns the fully-qualified name, e.g. rerun.archetypes.Points3D:rerun.components.Position3D#positions.

This is the default Display implementation for ComponentDescriptor.

pub fn short_name(&self) -> String

Returns the unqualified name, e.g. Points3D:Position3D#positions.

impl ComponentDescriptor

pub fn new(component_name: impl Into<ComponentName>) -> Self

pub fn with_archetype_name(self, archetype_name: ArchetypeName) -> Self

Unconditionally sets Self::archetype_name to the given one.

pub fn with_archetype_field_name( self, archetype_field_name: ArchetypeFieldName, ) -> Self

Unconditionally sets Self::archetype_field_name to the given one.

pub fn or_with_archetype_name( self, archetype_name: impl Fn() -> ArchetypeName, ) -> Self

Sets Self::archetype_name to the given one iff it’s not already set.

pub fn or_with_archetype_field_name( self, archetype_field_name: impl FnOnce() -> ArchetypeFieldName, ) -> Self

Sets Self::archetype_field_name to the given one iff it’s not already set.