Module re_types_core::loggable

Structs

ComponentName

The fully-qualified name of a Component, e.g. rerun.components.Position2D.

DatatypeName

The fully-qualified name of a [Datatype], e.g. rerun.datatypes.Vec2D.

Traits

Component

A Component describes semantic data that can be used by any number of Archetypes.

Loggable

A Loggable represents a single instance in an array of loggable data.

Type Aliases

ComponentNameSet

UnorderedComponentNameSet